backend/requirements.txt
anthropic>=0.25.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.7.0
websockets>=12.0
